Extend forecast to 10 days and include nearly all xml tags
This is a very large change but SHOULD retain backwards compatibility.
Due to WebParser limitations the regexp parsing is now done in chunks
inside each .inc file with the GrabWeather.inc only fetching the xml
page.

This extends the forecast to the maximum supported value of 10 days ands
implements a measure for virtually every available tag.